#109 moved these test from prev module
