+++ batching-core/src/test/java/com/flipkart/batchingcore/data/DataTest.java
@@ -22,25 +22,46 @@
  *  THE SOFTWARE.
  */
 
-package com.flipkart.batching.exception;
+package com.flipkart.batchingcore.data;
+
+import com.flipkart.batchingcore.Data;
 
 import junit.framework.Assert;
 
 import org.junit.Test;
 
-import java.io.IOException;
+public class DataTest {
 
-/**
- * Test for {@link SerializeException}
- */
-public class SerializeExceptionTest {
 
     /**
-     * Test to verify {@link SerializeException#getRealException()} returns error
+     * Test to verify {@link Data#equals(Object)}
      */
     @Test
-    public void testSerializeException() {
-        SerializeException serializeException = new SerializeException(new IOException());
-        Assert.assertNotNull(serializeException.getRealException());
+    public void testEqualsData() {
+        Data data = new EventData();
+        Assert.assertTrue(!data.equals("e"));
+    }
+
+    /**
+     * Test to verify {@link TagData}
+     */
+    @Test
+    public void testTagData() {
+        Tag AD_TAG = new Tag("ADS");
+        TagData tagData = new TagData(AD_TAG);
+        Assert.assertTrue(tagData.getTag() == AD_TAG);
+    }
+
+    /**
+     * Test to verify the equals method in {@link TagData}
+     */
+    @Test
+    public void testTagEqualsData() {
+        Tag AD_TAG = new Tag("ADS");
+        Tag DEBUG_TAG = new Tag("DEBUG");
+        TagData adsTagData = new TagData(AD_TAG);
+        TagData debugTagData = new TagData(DEBUG_TAG);
+        Assert.assertTrue(!adsTagData.equals(debugTagData));
+        Assert.assertTrue(!adsTagData.equals(""));
     }
 }
