+++ batching-core/src/test/java/com/flipkart/batchingcore/exception/DeserializeExceptionTest.java
@@ -22,46 +22,25 @@
  *  THE SOFTWARE.
  */
 
-package com.flipkart.batching.data;
+package com.flipkart.batchingcore.exception;
 
 import junit.framework.Assert;
 
 import org.junit.Test;
 
-public class TagTest {
+import java.io.IOException;
+
+/**
+ * Test for {@link DeserializeException}
+ */
+public class DeserializeExceptionTest {
 
     /**
-     * Test to verify {@link Tag#getId()}
+     * Test to verify {@link DeserializeException}
      */
     @Test
-    public void setTag() {
-        Tag tag = new Tag("Ads");
-        Assert.assertTrue(tag.getId().equals("Ads"));
-
-        Tag tag1 = new Tag("Ads");
-        tag1.setId("Business");
-        Assert.assertTrue(tag1.getId().equals("Business"));
-    }
-
-    /**
-     * Test to verify {@link Tag#equals(Object)}
-     */
-    @Test
-    public void testEquals() {
-        Tag tag = new Tag("Ads");
-        Tag tag1 = new Tag("Business");
-
-        String string = "123";
-        Assert.assertTrue(!tag.equals(tag1));
-        Assert.assertTrue(!tag.equals(string));
-    }
-
-    /**
-     * Test to verify {@link Tag#hashCode()}
-     */
-    @Test
-    public void testHashCode() {
-        Tag tag = new Tag("Ads");
-        Assert.assertTrue(tag.hashCode() == tag.getId().hashCode());
+    public void testDeserializeException() {
+        DeserializeException deserializeException = new DeserializeException(new IOException());
+        Assert.assertNotNull(deserializeException.getRealException());
     }
 }
